did separation of concerns for data fetching <-> main section. further action required
